Create minimal flask app with first 2 scenes
The flask app is contained within app.py and very basically has get requests for the home screen and the 2 scenes.

For the html, create a class named "opt." opt are options that the user can select. Use javascript to allow the user to select between opts using the up and down arrow keys. Use this on the home screen and when selecting names in scene1. When a user presses enter on an opt that is also of class "submit," the parent form of the opt will be submitted, allowing for links and for regular form submission.

Use a placeholder donkey image for all images for now.