Add navigation within each scene, with pages
The game only displays one page at a time, and focus can be switched between them with special opt p elements that also contain "navi" class, with either "navi-prev" or "navi-next" as additional classes. Add a two macros, navi and page to base.html. page creates a new page with the content of the variable passed in, either hidden or not hidden depending on the second argument. navi creates a navi opt with the class specified.

Move scripts into the scripts folder within the static resources folder.