Add navigation to a specific page and part of scene 2
If a nav opt also has the class "navi-page," submitting it will navigate to the page with the id contained in data-navi. Add an argument to base.page, id, which will be the id of the new page. Also, allow the content of the opt to be changed. Add This page navigation will be first used in scene2, where the user will navigate between pages of the market.

Add dashed lists.

Fix bug where opt navigation can go past the current page's limit.