Add hidden and disabled classes and scene 2 logic
Hidden is used to hide an opt, especially a navi. In base.js, do not allow hidden opts to be selected. Disabled opts can be selected but no action will be taken if one is submitted.

When the player sells all the items, enable and reveal the opt to continue to the dealership. Add scene 3 backbone.